requirements.txt
Flask==2.3.2
tensorflow==2.12.0
Pillow==9.2.0
numpy==1.23.5
